[AppKit] Move the rest of the enums to the API definition.
This makes it easier to include/exclude certain enums in Mac Catalyst (since
we can use [NoMacCatalyst] instead of littering the files with "#if
__MACCATALYST__").

There should be no API changes at all.
